Update to latest
This is the first official public release of MouseTracks!

After originally starting work on the project in 2017, then picking it up again at the beginning of 2025, I'm happy to finally say it's reached a point where it can be released.
All the core features are in place, with a lot of effort put into making sure the application runs as smoothly and as reliably as possible.

# Features
- Live activity tracking
- Per-application profiling
- High quality rendering
- Customisable artistic renders
- Multi-monitor support
- Interactive live GUI
- Import/export data
- Standalone application (no install required)
- Option to launch on startup